Feat: Refactored and cleaned bill component to use item map instead of item array
