+++ app/components/table/bill/item/atoms/delete.tsx
@@ -1,0 +1,7 @@
+export default function DeleteItem({
+    removeDelete,
+}: {
+    removeDelete: () => void;
+}) {
+    return <div>Delete</div>;
+}
